Moved some UI classes to different packages
