+++ composeApp/src/commonMain/kotlin/de/medizininformatikinitiative/medgraph/ui/searchengine/results/SearchResultsListViewModel.kt
@@ -1,4 +1,4 @@
-package de.medizininformatikinitiative.medgraph.ui.searchengine
+package de.medizininformatikinitiative.medgraph.ui.searchengine.results
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
